requirements.txt
Django==5.2.4
gunicorn==23.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
dj-database-url==2.1.0
whitenoise==6.8.1
django-cors-headers==4.3.1
pillow==10.4.0
